pof-ai-log/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
google-cloud-kms==2.21.4
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
